Move core navigation components to shared module
