+++ src/App.tsx
@@ -1,13 +1,12 @@
 import { lazy, Suspense } from "react";
-import CAGEDVisualizer from "./components/CAGEDVisualizer";
-import AppNavigation from "./components/AppNavigation";
+import CAGEDVisualizer from "@/systems/caged/components/CAGEDVisualizer";
+import { AppNavigation, LoadingFallback } from "@/shared/components";
 import { ThemeProvider } from "./contexts/ThemeContext";
 import { NavigationProvider } from "./contexts/NavigationContext";
 import { useNavigation } from "./hooks/useNavigation";
-import { LoadingFallback } from "./components/LoadingFallback";
 
 // Lazy load Quiz components for better initial bundle size
-const QuizPage = lazy(() => import("./components/QuizPage"));
+const QuizPage = lazy(() => import("@/systems/quiz/components/QuizPage"));
 
 function AppContent() {
   const { currentPage } = useNavigation();
