+++ src/shared/components/index.ts
@@ -3,3 +3,6 @@
  */
 
 export { default as FretboardDisplay } from './FretboardDisplay';
+export { default as AppNavigation } from './AppNavigation';
+export { default as ThemeToggle } from './ThemeToggle';
+export * from './LoadingFallback';
